move $.resolve to $.entity.resolve
