+++ packages/compiler/src/typekit/index.ts
@@ -20,7 +20,6 @@ export {
   OperationDescriptor,
   OperationKit,
   RecordKit,
-  ResolveKit,
   ScalarKit,
   TupleKit,
   TypeTypekit,
